% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis_scale.R
\name{mahalanobis_scale}
\alias{mahalanobis_scale}
\title{Perform Mahalanobis transformation.}
\usage{
mahalanobis_scale(X)
}
\arguments{
\item{X}{A data matrix.}
}
\value{
The transformed data matrix.
}
\description{
This function performs the Mahalanobis transformation on a data matrix.
}
\examples{
data(heptathlon)
mahalanobis_scale(heptathlon)

}
